+++ app/assets/stylesheets/application.css
@@ -103,25 +103,25 @@ h4 {
   margin-right: 1rem;
 }
 
-.buttons {
+.buttons,
+.post-comment {
   display: flex;
-
-  /* gap: 9rem; */
   justify-content: space-between;
   margin-right: 1rem;
   margin-left: 1rem;
 }
 
-.home-link {
-  border: 1px solid;
-  padding: 1rem;
-}
-
-.back-link {
+.home-link,
+.back-link,
+#like-button {
   border: 1px solid;
   padding: 1rem;
 }
 
 h1 {
   margin-left: 1rem;
+}
+
+form {
+  padding: 1rem;
 }
