+++ app/helpers/comments_helper.rb
@@ -1,0 +1,2 @@
+module CommentsHelper
+end
