+++ app/views/posts/index.html.erb
@@ -22,7 +22,7 @@
     <% @user.posts.paginate(page: params[:page], per_page: (params[:per_page] || 2).to_i).each_with_index do |post, index| %>
       <li class="list">
         <div class="post">
-          <%= link_to post_path(user_id: @user.id, id: post.id) do %>
+          <%= link_to user_post_path(user_id: @user.id, id: post.id) do %>
 <h4>Post #<%= (params[:page].to_i) * (params[:per_page] || 2).to_i + index + 1 %></h4>
           <% end %>
           <p><%= post.text %></p>
